+++ bootstrap/ssr/ssr-manifest.json
@@ -107,79 +107,25 @@
     "/build/assets/debug_ini-ac2PUNJ1.png",
     "/build/assets/cs50-CmGKmCZY.png"
   ],
-  "resources/js/Components/CommentaryForm.vue": [
-    "/build/assets/me-7NLqheFn.jpg",
-    "/build/assets/github-DJ4kBwNf.svg",
-    "/build/assets/linkedin-CsZAhWbA.svg",
-    "/build/assets/devcontainerfolder-DTN3fVqK.png",
-    "/build/assets/vsextension-BUuXiwhS.png",
-    "/build/assets/dockerfile-C_VCj5mP.png",
-    "/build/assets/debug-CtGYDOVf.png",
-    "/build/assets/laravel-welcome-page-ez6egKWo.png",
-    "/build/assets/laravel-welcome-Jetstream-BNoQpVBe.png",
-    "/build/assets/laravel-dashborad-C6SkwNc0.png",
-    "/build/assets/requirefields-CEQ8spP0.png",
-    "/build/assets/pest-output-DAWTyeAj.png",
-    "/build/assets/pest-error-output-DlXUmicb.png",
-    "/build/assets/pest-output-min-rUg6y2nq.png",
-    "/build/assets/debug_ini-ac2PUNJ1.png",
-    "/build/assets/cs50-CmGKmCZY.png"
-  ],
-  "resources/js/Components/CommentaryList.vue": [
-    "/build/assets/me-7NLqheFn.jpg",
-    "/build/assets/github-DJ4kBwNf.svg",
-    "/build/assets/linkedin-CsZAhWbA.svg",
-    "/build/assets/devcontainerfolder-DTN3fVqK.png",
-    "/build/assets/vsextension-BUuXiwhS.png",
-    "/build/assets/dockerfile-C_VCj5mP.png",
-    "/build/assets/debug-CtGYDOVf.png",
-    "/build/assets/laravel-welcome-page-ez6egKWo.png",
-    "/build/assets/laravel-welcome-Jetstream-BNoQpVBe.png",
-    "/build/assets/laravel-dashborad-C6SkwNc0.png",
-    "/build/assets/requirefields-CEQ8spP0.png",
-    "/build/assets/pest-output-DAWTyeAj.png",
-    "/build/assets/pest-error-output-DlXUmicb.png",
-    "/build/assets/pest-output-min-rUg6y2nq.png",
-    "/build/assets/debug_ini-ac2PUNJ1.png",
-    "/build/assets/cs50-CmGKmCZY.png"
-  ],
-  "resources/js/Components/CommentaryList.vue?vue&type=style&index=0&lang.css": [
-    "/build/assets/me-7NLqheFn.jpg",
-    "/build/assets/github-DJ4kBwNf.svg",
-    "/build/assets/linkedin-CsZAhWbA.svg",
-    "/build/assets/devcontainerfolder-DTN3fVqK.png",
-    "/build/assets/vsextension-BUuXiwhS.png",
-    "/build/assets/dockerfile-C_VCj5mP.png",
-    "/build/assets/debug-CtGYDOVf.png",
-    "/build/assets/laravel-welcome-page-ez6egKWo.png",
-    "/build/assets/laravel-welcome-Jetstream-BNoQpVBe.png",
-    "/build/assets/laravel-dashborad-C6SkwNc0.png",
-    "/build/assets/requirefields-CEQ8spP0.png",
-    "/build/assets/pest-output-DAWTyeAj.png",
-    "/build/assets/pest-error-output-DlXUmicb.png",
-    "/build/assets/pest-output-min-rUg6y2nq.png",
-    "/build/assets/debug_ini-ac2PUNJ1.png",
-    "/build/assets/cs50-CmGKmCZY.png"
-  ],
-  "resources/js/Components/CommentaryLogin.vue": [
-    "/build/assets/me-7NLqheFn.jpg",
-    "/build/assets/github-DJ4kBwNf.svg",
-    "/build/assets/linkedin-CsZAhWbA.svg",
-    "/build/assets/devcontainerfolder-DTN3fVqK.png",
-    "/build/assets/vsextension-BUuXiwhS.png",
-    "/build/assets/dockerfile-C_VCj5mP.png",
-    "/build/assets/debug-CtGYDOVf.png",
-    "/build/assets/laravel-welcome-page-ez6egKWo.png",
-    "/build/assets/laravel-welcome-Jetstream-BNoQpVBe.png",
-    "/build/assets/laravel-dashborad-C6SkwNc0.png",
-    "/build/assets/requirefields-CEQ8spP0.png",
-    "/build/assets/pest-output-DAWTyeAj.png",
-    "/build/assets/pest-error-output-DlXUmicb.png",
-    "/build/assets/pest-output-min-rUg6y2nq.png",
-    "/build/assets/debug_ini-ac2PUNJ1.png",
-    "/build/assets/cs50-CmGKmCZY.png"
-  ],
   "resources/js/Components/CommentarySection.vue": [
+    "/build/assets/me-7NLqheFn.jpg",
+    "/build/assets/github-DJ4kBwNf.svg",
+    "/build/assets/linkedin-CsZAhWbA.svg",
+    "/build/assets/devcontainerfolder-DTN3fVqK.png",
+    "/build/assets/vsextension-BUuXiwhS.png",
+    "/build/assets/dockerfile-C_VCj5mP.png",
+    "/build/assets/debug-CtGYDOVf.png",
+    "/build/assets/laravel-welcome-page-ez6egKWo.png",
+    "/build/assets/laravel-welcome-Jetstream-BNoQpVBe.png",
+    "/build/assets/laravel-dashborad-C6SkwNc0.png",
+    "/build/assets/requirefields-CEQ8spP0.png",
+    "/build/assets/pest-output-DAWTyeAj.png",
+    "/build/assets/pest-error-output-DlXUmicb.png",
+    "/build/assets/pest-output-min-rUg6y2nq.png",
+    "/build/assets/debug_ini-ac2PUNJ1.png",
+    "/build/assets/cs50-CmGKmCZY.png"
+  ],
+  "resources/js/Components/CommentarySection.vue?vue&type=style&index=0&lang.css": [
     "/build/assets/me-7NLqheFn.jpg",
     "/build/assets/github-DJ4kBwNf.svg",
     "/build/assets/linkedin-CsZAhWbA.svg",
